refactor: 회원 프로필 뷰 상태를 ViewState로 통합해 일관성 개선

+++ app/app/Component/MemberProfileInfo.swift
@@ -7,7 +7,7 @@ struct MemberProfileInfo: View {
     let gender: String
     let age: Int
     let bio: String
-    let likes: Int
+    let likes: Int64
     let distance: Double?
     
     var body: some View {
